+++ gvm/protocols/gmpv208/__init__.py
@@ -37,6 +37,7 @@ from gvm.protocols.gmpv208.entities.alerts import (
     get_alert_event_from_string,
     get_alert_method_from_string,
 )
+from gvm.protocols.gmpv208.entities.audits import AuditsMixin
 from gvm.protocols.gmpv208.entities.hosts import (
     HostsMixin,
 )
@@ -70,7 +71,7 @@ from gvm.protocols.gmpv208.entities.targets import (
     get_alive_test_from_string,
     TargetsMixin,
 )
-from gvm.protocols.gmpv208.entities.tasks import TaskMixin
+from gvm.protocols.gmpv208.entities.tasks import TasksMixin
 from gvm.protocols.gmpv208.entities.tls_certificates import TLSCertificateMixin
 from gvm.connections import GvmConnection
 
@@ -84,6 +85,7 @@ PROTOCOL_VERSION = (20, 8)
 class Gmp(
     GmpV208Mixin,
     AlertsMixin,
+    AuditsMixin,
     HostsMixin,
     NotesMixin,
     OperatingSystemsMixin,
@@ -92,7 +94,7 @@ class Gmp(
     ReportsMixin,
     ResultsMixin,
     TargetsMixin,
-    TaskMixin,
+    TasksMixin,
     TLSCertificateMixin,
     SecInfoMixin,
 ):